Location events refactoring

+++ Interfaces/Events/CombatStatistics.cs
@@ -5,27 +5,27 @@
     public class CombatStatistics
     {
         [JsonProperty("Bounties_Claimed")]
-        public long BountiesClaimed { get; set; }
+        public double BountiesClaimed { get; set; }
 
         [JsonProperty("Bounty_Hunting_Profit")]
-        public long BountyHuntingProfit { get; set; }
+        public double BountyHuntingProfit { get; set; }
 
         [JsonProperty("Combat_Bonds")]
-        public long CombatBonds { get; set; }
+        public double CombatBonds { get; set; }
 
         [JsonProperty("Combat_Bond_Profits")]
-        public long CombatBondProfits { get; set; }
+        public double CombatBondProfits { get; set; }
 
         [JsonProperty("Assassinations")]
-        public long Assassinations { get; set; }
+        public double Assassinations { get; set; }
 
         [JsonProperty("Assassination_Profits")]
-        public long AssassinationProfits { get; set; }
+        public double AssassinationProfits { get; set; }
 
         [JsonProperty("Highest_Single_Reward")]
-        public long HighestSingleReward { get; set; }
+        public double HighestSingleReward { get; set; }
 
         [JsonProperty("Skimmers_Killed")]
-        public long SkimmersKilled { get; set; }
+        public double SkimmersKilled { get; set; }
     }
 }
